refactor(Presentation): UI Build 최적화 - recordVM을 상위 객체로 이동 -> 데이터 fetch를 가리기 위함. - goal model 파리미터 Double로 수정 - coordinator hash swift6 대응 - Data 모델 Sendable 적용

+++ RecordManagment/Domain/Entitiy/Goal/GoalAchieve.swift
@@ -20,7 +20,7 @@ struct CurrentPeriodData: Decodable {
     let startDate: [Int]
     let endDate: [Int]
     let completedDays: Int
-    let achievementRate: Int
+    let achievementRate: Double
     let treeStage: String
     let isInProgress: Bool?
 }
@@ -32,7 +32,7 @@ struct RecentHistoryData: Decodable {
     let startDate: [Int]
     let endDate: [Int]
     let completedDays: Int
-    let achievementRate: Int
+    let achievementRate: Double
     let finalTreeStage: String
     let status: String
 }
